関数ごとにfunctionsのトリガーを分離
